+++ google/CHANGELOG.md
@@ -4,13 +4,14 @@ This file documents changes to the `workbench-google` library, including notes o
 
 ## 0.18
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.18-f074409"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.18-TRAVIS-REPLACE-ME"`
 
 ### Changed
 - Moved `SamModel`
 - add functions to access `DefaultObjectAccessControls` and `BucketAccesscontrols`
 - when getting service account treat 403 from google when error message is "Unable to extract resource containers." as service account does not exist see https://console.cloud.google.com/support/cases/detail/17978989?project=broad-dsde-prod
 - Add `GoogleKms` service and interpreter
+- Add `enableService` function to `GoogleProjectDAO`
 
 ## 0.17
 - upgrade cats to 1.4.0 and scala to 2.12.7